refactor: Rename XOR hash package
