+++ internal/xor/hash.go
@@ -1,4 +1,4 @@
-package hash
+package xor
 
 import "hash"
